Fix A10 feature in Firmata
In systemResetCallback()
Code will try to initialize all pins. On Atmega32U4, the last 11 pins, 18~29, are analog. So we do need to protect pin 26, which is also A8, D8, LIS3DH CS to make LIS3DH work. But make pin 28 PIN_MODE_IGNORE will disable analog function on A10. 28 is physical pin number of D8 on the chip, which has nothing to do with pinConfig.